% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_align.R
\name{CentroidBand}
\alias{CentroidBand}
\title{Centroid Band Extraction}
\usage{
pca_align(configuration = NULL)
}
\arguments{
\item{data}{A list of matrices of XYZ point clouds}
\item{pc}{Align along principal components before extracting}
\item{threshold}{Metric for band thickness relative to the point cloud scale}
}
\description{
Extracts centroid bands
}
\examples{
CentroidBand(data = pointclouds, pc = TRUE, threshold = 4)
}
\keyword{centroidband}
